Fix UInt64 Cast boolean conversion

+++ opc-ua-sdk/sdk-server/src/main/java/org/eclipse/milo/opcua/sdk/server/events/conversions/UInt64Conversions.java
@@ -29,7 +29,7 @@ final class UInt64Conversions {
 
   @NonNull
   static Boolean uInt64ToBoolean(@NonNull ULong ul) {
-    return ul.intValue() != 0;
+    return ul.longValue() != 0L;
   }
 
   @Nullable
